feat: show thread list sort order
